# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rooted Insight API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

/**
 * Status relative to reference range
 */
export type LabMarkerStatus =
  (typeof LabMarkerStatus)[keyof typeof LabMarkerStatus];

export const LabMarkerStatus = {
  normal: "normal",
  low: "low",
  high: "high",
  "critical-low": "critical-low",
  "critical-high": "critical-high",
} as const;

export interface LabMarker {
  /** Name of the lab marker (e.g., "Vitamin D", "TSH") */
  name: string;
  /** Numeric value of the marker */
  value: number;
  /** Unit of measurement (e.g., "ng/mL", "mIU/L") */
  unit: string;
  /** Lower bound of normal reference range */
  referenceRangeLow?: number | null;
  /** Upper bound of normal reference range */
  referenceRangeHigh?: number | null;
  /** Status relative to reference range */
  status: LabMarkerStatus;
}

export interface CreateLabResultRequest {
  /** Name of the lab panel/test (e.g., "Comprehensive Metabolic Panel") */
  testName: string;
  /** Date the test was taken */
  testDate: string;
  /** Name of the laboratory */
  labName?: string | null;
  /** Optional personal notes */
  notes?: string | null;
  /** Individual lab markers and values */
  markers: LabMarker[];
}

export type NaturalSuggestionCategory =
  (typeof NaturalSuggestionCategory)[keyof typeof NaturalSuggestionCategory];

export const NaturalSuggestionCategory = {
  supplement: "supplement",
  diet: "diet",
  lifestyle: "lifestyle",
  exercise: "exercise",
  sleep: "sleep",
  stress: "stress",
} as const;

export type NaturalSuggestionEvidenceLevel =
  (typeof NaturalSuggestionEvidenceLevel)[keyof typeof NaturalSuggestionEvidenceLevel];

export const NaturalSuggestionEvidenceLevel = {
  strong: "strong",
  moderate: "moderate",
  emerging: "emerging",
} as const;

export interface NaturalSuggestion {
  category: NaturalSuggestionCategory;
  title: string;
  description: string;
  evidenceLevel: NaturalSuggestionEvidenceLevel;
}

export interface MarkerInsight {
  markerName: string;
  /** Simple explanation of what this marker means */
  plainLanguageExplanation: string;
  /** Why this value matters */
  significance: string;
  suggestions: NaturalSuggestion[];
}

export interface LabAnalysis {
  id: number;
  labResultId: number;
  /** Overall summary in plain language */
  summary: string;
  markerInsights: MarkerInsight[];
  /** Medical disclaimer */
  disclaimer: string;
  createdAt: string;
}

export interface LabResult {
  id: number;
  testName: string;
  testDate: string;
  labName?: string | null;
  notes?: string | null;
  markers: LabMarker[];
  createdAt: string;
  analysis?: LabAnalysis | null;
}

export interface CreateMedicationRequest {
  /** Medication name */
  name: string;
  /** Dosage (e.g., "10mg") */
  dosage?: string | null;
  /** How often taken (e.g., "Once daily") */
  frequency?: string | null;
  /** What condition it is prescribed for */
  prescribedFor?: string | null;
  /** Prescribing healthcare provider */
  prescribedBy?: string | null;
  startDate?: string | null;
  notes?: string | null;
}

export interface Medication {
  id: number;
  name: string;
  dosage?: string | null;
  frequency?: string | null;
  prescribedFor?: string | null;
  prescribedBy?: string | null;
  startDate?: string | null;
  notes?: string | null;
  createdAt: string;
}