# lib/api-zod/src/generated/types/createLabResultRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rooted Insight API
 * OpenAPI spec version: 0.1.0
 */
import type { LabMarker } from "./labMarker";

export interface CreateLabResultRequest {
  /** Name of the lab panel/test (e.g., "Comprehensive Metabolic Panel") */
  testName: string;
  /** Date the test was taken */
  testDate: Date;
  /** Name of the laboratory */
  labName?: string | null;
  /** Optional personal notes */
  notes?: string | null;
  /** Individual lab markers and values */
  markers: LabMarker[];
}