# lib/api-zod/src/generated/types/createMedicationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rooted Insight API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMedicationRequest {
  /** Medication name */
  name: string;
  /** Dosage (e.g., "10mg") */
  dosage?: string | null;
  /** How often taken (e.g., "Once daily") */
  frequency?: string | null;
  /** What condition it is prescribed for */
  prescribedFor?: string | null;
  /** Prescribing healthcare provider */
  prescribedBy?: string | null;
  startDate?: Date | null;
  notes?: string | null;
}