# lib/api-zod/src/generated/types/labAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rooted Insight API
 * OpenAPI spec version: 0.1.0
 */
import type { MarkerInsight } from "./markerInsight";

export interface LabAnalysis {
  id: number;
  labResultId: number;
  /** Overall summary in plain language */
  summary: string;
  markerInsights: MarkerInsight[];
  /** Medical disclaimer */
  disclaimer: string;
  createdAt: Date;
}