# lib/api-zod/src/generated/types/labMarker.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rooted Insight API
 * OpenAPI spec version: 0.1.0
 */
import type { LabMarkerStatus } from "./labMarkerStatus";

export interface LabMarker {
  /** Name of the lab marker (e.g., "Vitamin D", "TSH") */
  name: string;
  /** Numeric value of the marker */
  value: number;
  /** Unit of measurement (e.g., "ng/mL", "mIU/L") */
  unit: string;
  /** Lower bound of normal reference range */
  referenceRangeLow?: number | null;
  /** Upper bound of normal reference range */
  referenceRangeHigh?: number | null;
  /** Status relative to reference range */
  status: LabMarkerStatus;
}