# lib/api-zod/src/generated/types/labResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rooted Insight API
 * OpenAPI spec version: 0.1.0
 */
import type { LabAnalysis } from "./labAnalysis";
import type { LabMarker } from "./labMarker";

export interface LabResult {
  id: number;
  testName: string;
  testDate: Date;
  labName?: string | null;
  notes?: string | null;
  markers: LabMarker[];
  createdAt: Date;
  analysis?: LabAnalysis | null;
}