# lib/api-zod/src/generated/types/markerInsight.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rooted Insight API
 * OpenAPI spec version: 0.1.0
 */
import type { NaturalSuggestion } from "./naturalSuggestion";

export interface MarkerInsight {
  markerName: string;
  /** Simple explanation of what this marker means */
  plainLanguageExplanation: string;
  /** Why this value matters */
  significance: string;
  suggestions: NaturalSuggestion[];
}