# lib/api-zod/src/generated/types/medication.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rooted Insight API
 * OpenAPI spec version: 0.1.0
 */

export interface Medication {
  id: number;
  name: string;
  dosage?: string | null;
  frequency?: string | null;
  prescribedFor?: string | null;
  prescribedBy?: string | null;
  startDate?: Date | null;
  notes?: string | null;
  createdAt: Date;
}