# lib/api-zod/src/generated/types/naturalSuggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rooted Insight API
 * OpenAPI spec version: 0.1.0
 */
import type { NaturalSuggestionCategory } from "./naturalSuggestionCategory";
import type { NaturalSuggestionEvidenceLevel } from "./naturalSuggestionEvidenceLevel";

export interface NaturalSuggestion {
  category: NaturalSuggestionCategory;
  title: string;
  description: string;
  evidenceLevel: NaturalSuggestionEvidenceLevel;
}